fix create project bug

+++ src/main/java/cnpm/doan/entity/User.java
@@ -10,7 +10,7 @@ public class User {
     @Column(name = "id")
     @SequenceGenerator(name = "user_seq", sequenceName = "user_seq", allocationSize = 1)
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "user_seq")
-    private int id;
+    private long id;
     @Column(name = "name")
     private String name;
     @Column(name = "password")
@@ -35,13 +35,13 @@ public class User {
 
     }
 
-    public User(int id, String email, String password) {
+    public User(long id, String email, String password) {
         this.id = id;
         this.email = email;
         this.password = password;
     }
 
-    public User(int id, String name, String password, String email, String address, String gender, Role roles, Department department) {
+    public User(long id, String name, String password, String email, String address, String gender, Role roles, Department department) {
         this.id = id;
         this.gender = gender;
         this.name = name;
@@ -100,11 +100,11 @@ public class User {
         this.department = department;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
